+++ src/app/api/posts/[id]/route.ts
@@ -1,7 +1,9 @@
 import { type NextRequest, NextResponse } from 'next/server';
+import { type SafeParseReturnType } from 'zod';
 import { type Post } from '@prisma/client';
-import { readPost } from '@/actions';
-import { prisma } from '@/lib/db';
+import { readPost, updatePost } from '@/lib/actions';
+import { postSchema } from '@/lib/schemas';
+import { type PostSchema } from '@/lib/types';
 
 /**
  * TODOs
@@ -31,4 +33,34 @@ const GET = async (
   });
 };
 
-export { GET };
+const PUT = async (
+  request: NextRequest
+): Promise<
+  NextResponse<{
+    data: Post | null;
+    errors: { [key: string]: string[] } | null;
+    success: boolean;
+  }>
+> => {
+  const payload: unknown = await request.json();
+  const parsedPayload: SafeParseReturnType<PostSchema, PostSchema> =
+    postSchema.safeParse(payload);
+
+  if (!parsedPayload.success) {
+    return NextResponse.json({
+      data: null,
+      errors: parsedPayload.error?.flatten().fieldErrors,
+      success: false,
+    });
+  }
+
+  const post: Post | null = await updatePost(parsedPayload.data);
+
+  return NextResponse.json({
+    data: post,
+    errors: null,
+    success: true,
+  });
+};
+
+export { GET, PUT };
